requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
jupyter==1.1.1
notebook==7.4.4
pandas==2.3.1
numpy==2.2.6
matplotlib==3.10.5
seaborn==0.13.2
scikit-learn==1.7.1
opencv-python==4.12.0.88
pillow==11.3.0
ultralytics==8.0.196
torch>=1.9.0
torchvision>=0.10.0 